submissions/2026-05-04__001_synthetic_mine_throughput__claude-code__claude-opus-4-7__shipoftheseus-v0914-cold01/requirements.txt
simpy>=4.0
numpy>=1.24
pandas>=2.0
scipy>=1.10
networkx>=3.0
pyyaml>=6.0
matplotlib>=3.7
pytest>=7.0
